Fix batch split-K tests.
We should not expect to re-use prologue smem when we have batch &
split-K, because in that case the batch loop will exist in the kernel so
there will be no smem re-use in those cases. Note that this _should_ be
possible but requires a block sync at the end of the batch loop and we
currently do not support that pattern. See #1899