Rename `PredicateAnalyzer` -> `ProducerConsumerPairAnalyzer` (#1912)
Right now, we have `PredicateChcker` and `PredicateAnalyzer`. From their
name, we can not tell which is doing what. After reading the code, we
can tell that, `PredicateChcker` is the global entry point for predicate
elimination, and `PredicateAnalyzer` is a helper class for
`predicateProducerConsumerPair`, which is only one condition of
`PredicateChcker`:
https://github.com/NVIDIA/Fuser/blob/638bfe00a29f582a337042c04a63c13410b8b17f/csrc/device_lower/analysis/predicate_elimination.cpp#L222
So I do this rename to indicate this fact.