Multidevice: Avoid compiling when allocating (#1900)
This patch accelerates the multidevice allocation which was the main
overhead in our runtime.

Indeed, previously we needed to compile and run a Fusion just for
allocation. This path gets rid of this step by simply computing the size
of the allocation and allocating directly.

Replaces https://github.com/NVIDIA/Fuser/pull/1854

possible TODO (not sure it's worth it): cache the compute of the
allocation size depending on the input's sizes.