Sharded tensor metadata (#1932)
Redo of: https://github.com/NVIDIA/Fuser/pull/1897

Fixes an inconsistency in tensor metadata. We previously saved the
unsharded sizes in the logical size, but saved the sharded strides in
the logical strides. This is made consistent: the sharded sizes and
sharded strides are stored in tensor metadata.
Note, the TensorView's IterDomains still refer to the unsharded extents.
In index computation, we ensure sharded axes are treated as having an
extent of 1.

Tests that computation with non-outermost sharded indices is lowered and
executes properly.