TMA predicate elimination (#1933)
This PR modifies the tests `TMALdstTest.LoadCompleteTensor*` and
`TMALdstTest.StoreCompleteTensor*`. In the past, the tensors were not
scheduled at all (except for swizzling support). We just took whatever
in the rFactor domain and parallelize them as `Bulk`. Now, we are
actually creating a tile `[M, N, ...]` -> `[M/tile_sizes[0] *
N/tile_sizes[1] * ..., tile_sizes[0] * tile_sizes[1] * ...]`.

The limitation that we can only do whole tensor copy still exist, but
this time it is a dynamic shape tensor that happens to be a single tile,
instead of having a static single tile tensor. Due to the change of
schedule, our system will try to create predicates for the TMA
expression, which requires me to change the predicate elimination pass
to eliminate these predicates. For the specific case of complete tensor
load/store, the TMA expression should not be predicated, and I am adding
a check for that in the test.