Check for SliceOp before inserting casts (#1936)
We already check for select-based ops because they must be the first op
in the segment. This PR just adds SliceOp to the list of types to check.

This current approach is flawed, since it will cause some segments to
not cast at all, leading to increased IO. Instead, we should probably
forward the casts through incompatible ops instead of skipping inserting
casts altogether.

Fixes #1953.