matmul via ATen: converting enable option to disable (#1927)
Our default behavior for the Matmul API in design (in the stacked PR:
https://github.com/NVIDIA/Fuser/pull/1927#issue-2184766755 ) is to
evaluate via ATen/Expression Evaluator (EE). To make the stacked a bit
cleaner and to express the default behavior better, we have a disable
flag to to stop execution via EE. Thus, in the usual case it'll execute
via EE.

I changed the constructor of our NVFuserTests to disable this options so
that we can keep testing Fuser's matmul scheduling and lowering.