A Python API for Matmul (#1878)
We add a new Python API in Fuser, which will accept tensors of the shape
(M,K), (N,K). Currently we will execute the Matmul using the Aten
expression evaluator, but will move to Fuser's scheduling and lowering
in the future.
This API will only allow tensors with the shape (M,K) (K,N), but the
tensors may be strided, thus allowing different layouts.