Add kir::Allocate::resetsToZero()
This will allow us to incrementally enable reuse of zeroed memory
(see #1984) on a case-by-case basis. To do so, when creating an
`Allocate` node, if it is guaranteed to be "cleaned up" (i.e. reset to
all zero values) by the end of the kernel, then you can pass `true` for
the `resets_to_zero` argument in the `kir::Allocate` constructor. This
will automatically enable reuse of zeroed memory for this buffer and
avoid a memset call, regardless of whether
`EnableOption::ReuseZeroedMemory` is enabled or not. Eventually this
will let us get rid of that option entirely, once all of our global
memory semaphores are guaranteed to be reset.