Hack to ensure cp.async is waited before smem reuse (#2001)
This is a work-around for #2000.

It seems to address the issue in the only current use case for smem
reuse: matmul with `params.use_smem_epilogue == true`. It is not ideal:
for example it will insert a `cp.async.wait_all` instruction even if
circular buffering is not used in the kernel.

Fixes #1996 but since this is a hack, I will not mark #2000 as fixed
yet.