Introduce Features
This change introduces Feature, which is an enum class that will
eventually replace EnableOption and DisableOption. Features can default
to enabled or disabled. We can collect them together into a packed
structure called FeatureSet, which is a thin wrapper around std::bitset
that can be re-used with any enum containing EndOfEnum as its largest
value.

This is meant to enable us to move away from using global state to
manage options, and instead include the set of enabled features as
attributes in the FusionExecutorCache stack.