Simplify resize extent (#2017)
I noticed the extent of a slice output IterDomain is not simplified as
intended. Currently we do pattern matching of `a - b`, but we actually
get `a + (-b)` due to `SimplifyingIrBuilder`. This is a quick fix to
make sure the latter pattern is also detected. For example:

Before: `i0 + ( ( fmax(0, ( fmin(i0, 1) )) ) + ( -i0 ) )`
After: `fmax(0, ( fmin(i0, 1) ))`