Multidevice staged reduction with composed manual and auto-scheduling (#1853)
This PR implements the first composition of **manual interdevice
scheduling** and **automatic intradevice scheduling.**

Besides being useful in itself, we also hope that this could serve as an
example to help the team modify other schedulers to support
inter-/intra- device scheduling in a wider range of scenario.

## What

Let us focus on the test implemented in
`test/test_multidevice_pipeline.cpp`. There, we consider a simple fusion
which expresses a multi-stage reduction:

```
  int num_devices = communicator->size();
  
  TensorView* tv0 = makeConcreteTensor([num_devices, 8, 32]);
  TensorView* tv1 = sum(tv0, {2});
  TensorView* tv_out = sum(tv1, {0});
  fusion->addInput(tv0);
  fusion->addOutput(tv_out);
```

_(Note that, for now, we only support reduction which are either
non-resharding (i.e. not reducing a sharded dimension) or resharding but
affecting only one axis. On the contrary, as far as intra-device
scheduling is concerned, Fuser typically requests consecutive reductions
to be merged. This inconsistency is only superficial, as we can always
translate one case to another by splitting/merging consecutive RedOps)_

We decide to manually shard the tensors on their outmost axis:

```
  std::vector<int64_t> devices(num_devices);
  std::iota(devices.begin(), devices.end(), 0);
  DeviceMesh mesh(devices);
  for (auto tv : {tv0, tv1, tv_out}) {
    tv->setDeviceMesh(mesh);
  }
  for (auto tv : {tv0, tv1}) {
    tv->axis(0)->parallelize(ParallelType::DIDx);
  }
```

This constitutes our multidevice scheduling.

Then, we test composing this interdevice scheduling with different modes
of intra-device scheduling:

- No scheduling: which is the only mode supported before this patch
- Manual scheduling: where we try a typical fully featured reduction
scheduling. Supporting this mode only required very minor changes.
- Automatic scheduling using `scheduleReduction` and `FusionExecutor`:
this required careful modification of the reduction scheduler.
- Fully automated segmentation+scheduling using `FusionExecutorCache`:
which is our ultimate target.

## How

The core modifications of the reduction scheduler can be simply
explained as follow. Before this patch, the scheduler would first
transform the underlying TensorDomain into the canonical form

```
{all_iteration_domains_merged, all_reduction_domains_merged}
```

Note that `all_iteration_domains_merged` can possibly be empty. We
modify this canonical form to be

```
{possibly_one_sharded_outmost_axis, all_iteration_domains_merged, all_reduction_domains_merged}
```

Then, we only need to possibly offset the axis we operate upon, in order
to leave untouch the outmost sharded axis.

We made sure that the behavior of the the scheduler remains unchanged in
the absence of multidevice scheduling.

### Other  modifications:

- MultiDeviceExecutor: handle passing the launch params
- Segmenter: dont' add NamedScalar to group's input
- multidevice Test: add the option to validate using prescribed tensors.