IdModel: Always require promotion to loop mapped domains (#2059)
This PR changes how Step 2 of the loop promotion analysis finds
promotion domains. Loop promotion should be done to domains that are in
the same loop group, but I thought this could be achieved by just
checking the mapping for the final results at Step 5 and it wouldn't
matter for intermediate results of promotion generated at Step 2. While
that's probably the case, it also introduces some unexpected domain
mappings and I found they were really confusing while working with some
complex fusions. Besides, as suggested by @zasdfgbnm, the initial
decision doesn't seem to buy much but just adds some extra complexity.

The requirement flag is now removed. The change to the promotion code is
pretty minimal, but unfortunately, it meant many of the promotion
reference results needed to be updated for test validation. The design
doc was also updated.

For a concrete example, see
`IdModelTest.LoopPromotionPromoteToSameLoopGroup` and its corresponding
section in the design doc.

Closes #2027