Refactor creation and update of `FusionHeuristics` in `FusionKernelRuntime` (#1649)
This PR seeks to avoid accumulating all statements in the segmenter's
complete fusion. Instead of pseudo-segmenting the complete fusion, we
create a separate fusion for each segmented group. Any new statements
are stored in the segmented group's fusion. For `test_issue1277`,
serialization handles ~147K values and ~71K expressions with current
ToT. With this PR, we only serialize 28K values and 7K expressions.

`SegmentedFusion::makeInitialHeuristic` creates the heuristic for each
segmented group. It uses the `FusionSegmentGuard` to pseudo-segment the
complete fusion. The values and expressions created when making the
heuristics are all stored in the same fusion. For complex fusions with
several segments, a significant number of statements can be accumulated.

`test_issue1277` in `test_python_frontend.py` is a pathological case.
The concretized fusion contains 1.5K values and 300 expressions. After
creating the heuristics, the complete fusion has 10K values and 5K
expressions. In `FusionKernelRuntime::compileKernel`, we create a
segmented fusion that copies all the values and expressions from the
complete fusion. For this test, segmentation creates 14 segmented groups 
so serialization handles ~147K values and ~71K expressions.