Matmul heuristic plugin interface (#2049)
This change introduces a plugin interface for user-provided matmul
heuristics. To use it, set the environment variable
`$NVFUSER_MATMUL_HEURISTIC_PLUGIN` to the location of a dynamic library
to use for matmul heuristic computation (this variable can also be the
base filename of a library in `$LD_LIBRARY_PATH`). See
`examples/matmul_heuristic_plugin` for a simple example.

TODO:
- build example plugin as part of CI