Print exception messages during parallel compile
This just creates a string that is written to (guarded by mutexes) when
a compile job fails with the error message and "Segmentation group N: ".
That string is then appended to the exception in the main thread which
the user sees. This means we will see actual error messages instead of
the hint to use `NVFUSER_DISABLE=parallel_compile`.