Check expected smem size before device limit (#2115)
This just checks that our smem usage matches the expected amount that
was optionally set during scheduling before checking that it falls
within device limits. This helps debugging in some cases since it can
tell us that we are overrunning our smem because of misestimation or
because of a heuristic decision that misused a correct estimate. This
was helpful in discovering #2111.