Matmul vectorization heuristic (#2120)
This introduces `supported_vec_size`, a struct attribute of
`MatmulParams` that indicates the maximum vectorization supported by
operands A and B as well as epilogue tensors (which will all be
vectorized at the same width).

Note that we expect a temporary performance regression since we hardcode
vec_size of 4 for the output, which is suboptimal for half-precision
outputs. This will be remedied when we start inferring the proper
vectorization (see #2105).