Fix deserialization issue.
As @rdspring1 pointed out, the issue is that I was only
::computeArgs()ing along the main path, not the deserialization
path. This is because the deserialization path sets up the
ExecutorEntry without using initializeExecutorEntry, and so our
entry_initialized boolean never got set.

The fix is simple and arguably better even for the main path:
instead of checking whether a bool says we're initialized, and
relying on that bool being set in all the proper places, check
instead to see if the `args` we are interested in are *actually*
initialized.