Keep actual tensors in `getMaybeHeuristicsFor` (#2123)
In #1649, a change to `FusionKernelRuntime::getMaybeHeuristicsFor` was
made which replaced all inputs with "metadata-only" tensors. These
tensors have the same size, stride, and dtype as the actual inputs, but
their data pointer is always `nullptr` and they do not allocate any
memory. This interferes with `runtime_info.ptrOf` as seen in #2118.
Instead, in this PR I revert that part of #1649. The `ArgumentManager`
that holds the actual `at::Tensor`s is destroyed at the end of
`getMaybeHeuristicsFor`, as is the `PrecomputedValues`. I don't think
there are any dangling references to the actual tensors in the args, so
I _think_ this should not leak device memory: cc @rdspring1 to confirm.

See the included passing tests, which fail on main.

Fixes #2118.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>