Rewrite argument building to avoid allocs. (#2136)
This PR attempts to pre-allocate space for `Tensor` arguments, as
considered in #2135. The key insight is that, while tensor shapes are
dynamic, the rank of a tensor is not. This allows us to pre-build the
memory space for passing a kernel to the GPU, avoiding numerous
allocations on subsequent calls.

For the 4 kernels mentioned in #2135, this improves launch latency by
4.4%, 7.5%, 4.4%, and 16%, respectively. It should be an
across-the-board improvement (i.e. no regressions). For comparison to
the case highlighted in #2135, the overall `GetArgsBuffers` call drops
from 12.147us to 9.512us (~20%):

![step 3 45 493
ms](https://github.com/NVIDIA/Fuser/assets/60981/25354504-f047-4c7c-a3c3-38f0930d8aca)

Whether or not this improvement matters depends significantly on the
network itself. In many training scenarios, we can already launch
quickly enough that the subsequent kernel is enqueued before a current
kernel finishes executing. In these scenarios, improving launch latency
does not improve overall performance.

After this patch, the ExpressionEvaluator in GetArgsBuffers takes
0.604 + 0.459 + 0.457 + 4.276 + 0.609 + 0.353 == 6.758us, or ~71% of the
9.512us GetArgsBuffers time. This means we are starting to hit
diminishing returns from addressing anything sans ExpressionEvaluator.

Fixes #2135.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>