Fix printed PTX for hopper mma (#2194)
Printed code:
```diff
4,5c4,5
<     "  setp.ne.b32 p0, %5, 0;\n"
<     "  wgmma.mma_async.sync.aligned.m64n8k16.f32.bf16.bf16 {%0, %1, %2, %3}, {%4, %5, %6, %7}, %8, p0, %9, %10, %11;\n"
---
>     "  setp.ne.b32 p0, %9, 0;\n"
>     "  wgmma.mma_async.sync.aligned.m64n8k16.f32.bf16.bf16 {%0, %1, %2, %3}, {%4, %5, %6, %7}, %8, p0, %10, %11, %12;\n"
```

As a result of this bug, the `transB` of the RS variant of mmas is
always `1`, and the `transA` of the SS variant of mmas is also always 1,
and the `transB` is `transA`. Because previously the mma tests are
green, this means that their schedules are wrong but consistent with
this wrong PTX, and we need to fix them. This is left as a future task.