[WIP] Check that matmul epilogue contains only pointwise ops
This PR restricts the accepted matmul segments for the nvfuser matmul
scheduler to only those containing pointwise epilogues. Additionally, it
rules out cases for which we cannot yet reliably determine epilogue input
vectorization due to transposes (TODO, see #2169).

Note that this check can be lifted when more epilogue cases are
supported, e.g. #2213.

Fixes #2167.

This is stacked on #2175 and follow-up PR to that introducing LinearOp
because currently segmentation fails for matmuls unless the complete
fusion can be scheduled (see #1707). The MatmulOp and LinearOp IR nodes
remove the need to inspect operand producer branches, so segmentation
should work fine once that work is merged. This PR will be marked as
draft until then.