Add test that external matmul plugin is loaded properly (#2229)
This adds a C++ test that checks that an external matmul heuristic
plugin is loaded. This is useful in automated jobs where we build a
matmul heuristic plugin and run some benchmarks. Without this test it is
possible to improperly prepare the environment and wind up running
benchmarks with the default core heuristics. Instead we can first run
this test and fail the job early if it does not succeed. The test is
disabled by default so that it does not cause default builds to fail. To
run it use a command sequence like this:
```
export NVFUSER_MATMUL_HEURISTIC_PLUGIN=/path/to/plugin.so
build/test_matmul --gtest_also_run_disabled_tests --gtest_filter='MatmulSchedulerTest.DISABLED_RequireExternalPlugin`
```