IdModel: Step 5 of the loop promotion analysis (#2220)
This is the final step of the loop promotion analysis. The promotion map
is almost completed at Step 3, but some partially inlined domains need
one more propagation, which is done by Step 4 and Step 5. Step 5 is
mostly just a repeat of Step 3.

This basically concludes the loop promotion analysis, although there are
a couple of issues that were found while working on indexing (#2218).
Those issues will be addressed as further follow-up PRs.

- Step 1: #1650 
- Step 2: #1777 
- Step 3: #1830 
- Step 4: #2003