ensure coalesced gmem access when each thread has 32 bytes (#2203)
In iter grouped grid reduction,threads with `TIDy == 0` in each block,
write block reduction results to global memory for final grid reduction.
When vectorized by 8, each thread writes 8 x sizeof(fp32) = 32 bytes,
which exceeds the max vect write width of 16 bytes. In current main
branch, the write is split into 2 vectoried writes without changing the
data layout in global memory. It leads to a problem similar to using
AoS. As shown in the figure, to serve these 256 bytes of data, the main
branch needs 4 transactions but if we change to the bottom version, only
needs 2 transactions.

![image](https://github.com/NVIDIA/Fuser/assets/116412316/acf911af-b96f-49a0-bb78-dba2248e4524)

After this change, the gmem store transaction is reduced but **no
apparent influence on performance** since the memory bandwidth is
dominated by the loading of the input tensor. The cache also helps the
transaction in the approach used in the current main branch. ncu shows
the store transactions is reduced by half.

![image](https://github.com/NVIDIA/Fuser/assets/116412316/f0710c82-c850-481c-8473-ba1a3d025941)

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>
Co-authored-by: jjsjann123 <jiej@nvidia.com>