ensure coalesced access to shared memory in block reduction (#2204)
**Issue:** smem_offset is the offset into shared memory for the current
thread. It is calculated as:
` offset = reduction_idx * reduction_size + reduction_tid`.
For outer reduction where TIDy is in the reduction dimension and TIDx is
in the iteration dimension and TIDz is not used.
We have `reduction_tid = TIDy` and `reduction_idx = TIDx`. The offset
leads to stride access to shared memory since it equivalent to: `offset
= TIDx * blockDim.y + TIDy`
**Fix:** To avoid this, we should use the offset based on the indexing
of threads within a block.
`offset= threadIdx.x + threadIdx.y * blockDim.x + threadIdx.z *
blockDim.x * blockDim.y;`
After this change, the index of the reduction peer element is no longer
a simple offset with a stride of 1, it depends on the reduction dim. See
code change for details.

**Results:** 
(1) removes bank conflicts. baseline is main branch, 99% of bank
conflicts are removed, smem access efficiency
 `requests  /  wave fronts= 16000 / 16189 = 99%`

![image](https://github.com/NVIDIA/Fuser/assets/116412316/130dbc82-f09b-410d-a573-ee2105ffb68a)
(2) This fix achieved 10% performance increase for outer reduction with
small inputs compared with main branch. For large inputs, the iter
grouped version is used, will be addressed in another PR.

![image](https://github.com/NVIDIA/Fuser/assets/116412316/8fdcaf6d-f3c8-4346-82e9-263d66347db5)