Generate a DOT representation of fusion transform. (#2251)
This is a simple debugging utility that I started using for IdModel.
Thought it may be of useful in general.

I'm not a Graphviz expert but just tried a bunch of things to make it
the output look reasonably good to me. Any suggestions would be
appreciated.

Here's an example output. The fusion is from one of the reshape shmoo
tests.


![transform](https://github.com/NVIDIA/Fuser/assets/363169/ab34c8c5-eff9-47c2-842e-3018627b8837)