Shortest path between ValGroups in ValGraph (#2270)
(Extracted from #2238)

Traversal for finding the shortest path from ValGroups to another
ValGroups. The algorithm is based on the standard BFS traversal,
however, since ValGraph is not an undirected graph, the dependencies of
ValGroups and ExprGroups need to be
satisfied. Specifically, when visiting an ExprGroup, either its inputs
or outputs must be visited before. Similarly, when visiting
a ValGroup, there must be at least one defining ExprGroup or one use
ExprGroup that is already visited.

The main use case is tensor indexing (#2238), where a typical traversal
would be from loop domains to allocation domains. Some indexing-specific
specialization would be needed, for example, dependencies with broadcast
domains can be ignored as their index is always just zero. The indexing
shortest-path traversal would be implemented by subclassing this class.

Please see the new unit tests for concrete examples.