Generalize CombineMulSum as MatmulPatterns (#2272)
This replaces the `CombineMulSum` class with `MatmulPattern` in the
Matmul scheduler. Additionally, we use these matmul patterns to
determine the problem layout, IterDomain roles, and TensorView roles.
The allocation domain is used to determine the problem layout. The
matmul scheduler is updated to reject segments whose input allocation
domains are non-trivial (until that is supported eg. by #2226).

Note that this does not add handling of `MatmulOp` and `LinearOp` in the
matmul scheduler. That will be done next in #2236 or similar.

---------

Co-authored-by: Priya Mishra <52657555+Priya2698@users.noreply.github.com>
Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>