Remove MatmulRole::OUTPUT_AUX (#2289)
We already have `MatmulRole::OUTPUT_D`. We populate both of these by
reverse sorting all outputs that are downstream of the `MmaOp` first by
number of uses then by `name()`. The first in that reverse sorted list
is `OUTPUT_D` and all others are `OUTPUT_AUX`. After this PR, there is
only the `OUTPUT_D` list but it might have more than one element: namely
the former `OUTPUT_AUX` tensors are now appended in the same order to
the `OUTPUT_D` list in the roles map.

I believe the main reason to reverse sort by uses is to find a reference
tensor for reverse propagating transforms to epilogue inputs here:
https://github.com/NVIDIA/Fuser/blob/0892405e5396820ffc4a300cba916cd6cb65e6be/csrc/scheduler/matmul.cpp#L660-L672
There should be no change to this section since we have not changed the
first element of the `OUTPUT_D` vector.