Loop promotion fix for loop groups with view rf domains (#2298)
Fixes #2296 

In addition to stopping at view rfactor domains as mentioned in the
issue page, I also put in a quick shortcut for the case where a view
rfactor domain is a terminal domain. In that case, it's guaranteed that
the domain is the right promotion domain. This is a shortcut and not
strictly necessary for the correctness of the analysis, but I think it's
simple enough to have as a quick optimization.