Set allocation domain of sharded tensor (#2271)
Sets allocation domain of sharded tensors during the pass
`propagateShardingsAndSetAllocationDomain`.
The two passes are merged in attempt to reduce the number of passes over
all expressions in the fusion.

Allocation domain is set to the tv's leaf domain. Since presegmentation
passes and scheduling occur after the sharding passes, the leaf domain
is identical to the rfact domain. After DID parallelization of the leaf
domain is allowed the leaf and rfactor domain will not be the same.

This will avoid issues such as
https://github.com/NVIDIA/Fuser/pull/2245#issuecomment-2111317055 and
allow the `AllocationDomainPass` presegmentation pass on for distributed
matmul tests