Translate MatmulOp and LinearOp (#2236)
The purpose of this PR is to enable the NVFuser matmul scheduler to
operate on the new `MatmulOp` an `LinearOp` nodes. That means the matmul
scheduler can optionally schedule segments if they are not supported by
ATen's `matmul` and `linear` functions.

Specifically, this PR:

1. Adds `MatmulOp` and `LinearOp` to the set of `MatmulPattern`s
detected. Previously only `MmaOp` and combined mul-sum patterns were
detected.
4. Enable translation of `MatmulOp` and `LinearOp` to fixed
broadcast+MmaOp patterns.
5. Introduces `EnableOption::FuseMatmul` and associated
`NVFUSER_ENABLE=fuse_matmul` option to enable the automatic scheduler to
accept matmul patterns consisting of a `MatmulOp` or `LinearOp`. By
default, the matmul scheduler will not accept segments containing
`MatmulOp` or `LinearOp` patterns, meaning all those nodes will be
computed using the `ExprEval` scheduler (ATen).

Not all cases that are supported by these IR nodes can be translated to
an `MmaOp`. In particular:
1. gemv cases where one operand is 1D are not supported
2. Cases with multiple batch dimensions.
3. Cases where "batch" dimensions must be added by unsqueezing one
dimension. Those new dimensions are indistinguishable from e.g. multiple
M dimensions which do not yet support.

These correspond to test cases in the two new tests. See the `TODO`
comments for descriptions of cases we plan to support but cannot yet
translate.

There is also a commented out test case for `LinearOp`: M=N=1 which
should not be translated to a `LinearOp` at all.

---------

Co-authored-by: Priya Mishra <52657555+Priya2698@users.noreply.github.com>