Remove LdMatrixTranspose type from LoadStoreOp types (#2315)
This PR modifies how we determine when we do a LdMatrix vs
LdMatrixTranspose.

We look at the (maybe) allocation domain of the producer and consumer of
the load store op do determine if a transpose is required. We get the
allocation domain of the consumer and check if the innermost dimension
maps to the innermost dimension of the allocation domain of the
producer. We transpose if it doesn't map.

---------

Co-authored-by: Gao, Xiang <qasdfgtyuiop@gmail.com>
Co-authored-by: Jacob Hinkle <1454944+jacobhinkle@users.noreply.github.com>